renames BigLabel.svelte to ToplineRow.svelte
